refactor(runtime): extract phase coordination module (#3057)
